% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CePa.R
\name{CePa}
\alias{CePa}
\title{Centrlity-based Pathway entrichment (CePa)}
\usage{
CePa(x, group, pathways, type, which = "proteins", edgeType = NULL,
  preparePaths = TRUE, norm.method = NULL, test.method = NULL,
  p.th = 0.05, logFC.th = 2, nperm = 1000, both.directions = TRUE,
  maxNodes = 150, minEdges = 0, commonTh = 2, filterSPIA = FALSE,
  convertTo = "none", convertBy = NULL)
}
\arguments{
\item{x}{An \code{ExpressionSet} object or a gene expression data matrix or count matrix, rows refer to genes, columns to samples}

\item{group}{Name or number of the phenoData column or a character vector or factor that contains required class assigments}

\item{pathways}{A list of pathways in a form from \code{graphite} package or created by \code{preparePathways()}}

\item{type}{Type of the input data, \code{"MA"} for microarray and \code{"RNASeq"} for RNA-Seq}

\item{which}{Character, which type of nodes is preserved in a pathway. Possible values are \code{"proteins"},\code{"metabolites"},\code{"mixed"}}

\item{edgeType}{Character, which type of edges is preserved in a pathway. If \code{NULL}, all edges are kept.}

\item{preparePaths}{Logical, by default the pathways are transformed with \code{preparePathways()}. Use \code{FALSE}, if you have done this transformation separately}

\item{norm.method}{Character, the method to normalize RNAseq data. If \code{NULL} then vst-normalization is performed. Possible values are: \code{"edgeR", "vst", "rLog", "none"}}

\item{test.method}{Character, the method for differentiall expression analysis of RNAseq data. If \code{NULL} then \code{"voomlimma"} is used. Possible values are: \code{"DESeq2", "voomlimma", "vstlimma", "edgeR"}.}

\item{p.th}{Numeric, threshold for p-values of tests for differential expression of genes. Use \code{1} if you don't want any threshold to be applied}

\item{logFC.th}{Numeric, threshold for log fold-change of a gene to identify the gene as differentially expressed. Use negative if you don't want any threshold to be applied}

\item{nperm}{Numeric, number of permutations}

\item{both.directions, maxNodes, minEdges, commonTh, filterSPIA, convertTo, convertBy}{Arguments for the \code{preparePathways()}}
}
\value{
A list:
\item{res}{A matrix, each row refers to one pathway, each column to one centrality and the value is a p-value.}
\item{topo.sig}{A list of weights for genes (nodes) in individual pathways}
\item{degtest}{A numeric vector of gene-level differential expression statistics of all genes in the dataset}
}
\description{
The function runs CePa method on microarray or RNA-Seq data. The implementation includes the identification of differentially expressed genes and transformation of pathways' topologies to an appropriate form. Only the ORA version of the CePa method is implemented and covers centralities: equal-weight, in-degree, out-degree, in-reach, out-reach and betweenness.
}
\examples{
if (require(breastCancerVDX)) {
data("vdx")
pathways<-pathways("hsapiens","biocarta")[1:3]
MAdata<-Biobase::exprs(vdx)[,1:10]
rownames(MAdata)<-Biobase::fData(vdx)[,"Gene.symbol"]
MAdata<-MAdata[!duplicated(rownames(MAdata)),]

CePa(MAdata, Biobase::pData(vdx)[,"er"][1:10], pathways, type="MA", convertTo="SYMBOL")
}
}
\references{
Gu Z., Liu J., Cao K., Zhang J., Wang J.: Centrality-based pathway enrichment: a systematic approach for finding significant pathways dominated by key genes. BMC Systems Biology 2012, 6:56
}
\author{
Ivana Ihnatova
}
\keyword{htest}
