\name{ToPASeq-deprecated}
\alias{ToPASeq-deprecated}
\alias{AdjacencyMatrix2pathway}
\alias{graphNEL2pathway}
\alias{KEGG2pathway}
\title{Deprecated functions in package \sQuote{ToPASeq}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{ToPASeq} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{AdjacencyMatrix2pathway: \code{\link{AdjacencyMatrix2Pathway}}}
    \item{graphNEL2pathway: \code{\link{graphNEL2Pathway}}}
    \item{KEGG2pathway: \code{\link{KEGG2Pathway}}}

  }
}
