% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{prs}
\alias{prs}
\alias{prsWeights}
\alias{prsWeights}
\title{Pathway regulation score (PRS)}
\usage{
prs(de, all, pwys, nperm = 1000)

prsWeights(pwy, de, all)
}
\arguments{
\item{de}{A named numeric vector containing log2 fold-changes of the 
differentially expressed genes. Recommended names are Entrez gene IDs.}

\item{all}{A character vector with the gene IDs in the reference set. If the
data was obtained from a gene expression experiment, this set will
contain all genes measured in the  experiment. This vector should contain 
*all* names of the \code{de} argument.}

\item{pwys}{A \code{linkS4class{PathwayList}} containing the pathways that 
should be analyzed for enrichment.}

\item{nperm}{Integer. Number of permutations.}

\item{pwy}{A \code{linkS4class{Pathway}} for which the weights should be
computed.}
}
\value{
A \code{data.frame} with normalized score and p-value for each 
pathway analyzed.
}
\description{
This function implements the PRS method to analyze pathway enrichment of gene
expression data. For PRS, a gene weight correspond to the number of 
downstream differentially expressed genes.
}
\examples{

# pathways
library(graphite)
pwys <- pathways("hsapiens","kegg")[1:10]

# expression data
all <- nodes(pwys[[1]])
nds <- sample(all, 30)
de <- setNames(rnorm(30), nds)

# executing PRS
prsWeights(pwys[[1]], de, all)
prs(de, all, pwys, nperm=100) 

}
\references{
Ibrahim et al. (2012) A topology-based score for pathway 
enrichment. J Comput Biol, 19(5):563-73.
}
\seealso{
\code{\link{pathways}}
}
\author{
Ivana Ihnatova
}
