% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_orig.r
\name{PRS_wrapper}
\alias{PRS_wrapper}
\title{Pathway Regulation Score (PRS)}
\usage{
PRS_wrapper(x, group, pathways, type, which = "proteins",
  edgeType = NULL, preparePaths = TRUE, norm.method = NULL,
  test.method = NULL, p.th = 0.05, logFC.th = 2, nperm = 1000,
  both.directions = TRUE, maxNodes = 150, minEdges = 0,
  commonTh = 2, filterSPIA = FALSE, convertTo = "none",
  convertBy = NULL)
}
\arguments{
\item{x}{An \code{ExpressionSet} object or a gene expression data matrix or count matrix, rows refer to genes, columns to samples}

\item{group}{Name or number of the phenoData column or a character vector or factor that contains required class assigments}

\item{pathways}{A list of pathways in a form from \code{graphite} package or created by \code{preparePathways()}}

\item{type}{Type of the input data, \code{"MA"} for microarray and \code{"RNASeq"} for RNA-Seq}

\item{which}{Character, which type of nodes is preserved in a pathway. Possible values are \code{"proteins"},\code{"metabolites"},\code{"mixed"}}

\item{edgeType}{Character, which type of edges is preserved in a pathway. If \code{NULL}, all edges are kept.}

\item{preparePaths}{Logical, by default the pathways are transformed with \code{preparePathways()}. Use \code{FALSE}, if you have done this transformation separately}

\item{norm.method}{Character, the method to normalize RNAseq data. If \code{NULL} then vst-normalization is performed. Possible values are: \code{"edgeR", "vst", "rLog", "none"}}

\item{test.method}{Character, the method for differentiall expression analysis of RNAseq data. If \code{NULL} then \code{"voomlimma"} is used. Possible values are: \code{"DESeq2", "voomlimma", "vstlimma", "edgeR"}. This analysis is needed only for the visualization.}

\item{p.th}{Numeric, threshold for p-values of tests for differential expression of genes. Use \code{1} if you don't want any threshold to be applied}

\item{logFC.th}{Numeric, threshold for log fold-change of a gene to identify the gene as differentially expressed. Use negative if you don't want any threshold to be applied}

\item{nperm}{Numeric, number of permutations}

\item{both.directions, maxNodes, minEdges, commonTh, filterSPIA, convertTo, convertBy}{Arguments for the \code{preparePathways()}}
}
\value{
A list:
\item{res}{A data frame with normalized score, p-value and FDR-adjusted p-value for each pathway}
\item{topo.sig}{A list with log fold-changes and number of downstream differentially expressed nodes for nodes of individual pathways}
\item{degtest}{A named vector of statistics from testing the differential expression of genes}
}
\description{
A function runs PRS method on a gene expression data matrix or count matrix and vector dividing samples into two groups and a set of pathways from \code{graphite} package. The PRS method (please see Reference for the details) was adapted to \code{graphite}'s graphs where each node is represented only by one gene.
}
\examples{
if (require(breastCancerVDX)) {
data("vdx")
pathways<-pathways("hsapiens","biocarta")[1:3]
MAdata<-Biobase::exprs(vdx)[,1:10]
rownames(MAdata)<-Biobase::fData(vdx)[,"Gene.symbol"]
MAdata<-MAdata[!duplicated(rownames(MAdata)),]

PRS_wrapper(MAdata, Biobase::pData(vdx)[,"er"][1:10], pathways, type="MA", convertTo="SYMBOL", logFC.th=-1, nperm=100)
}
}
\references{
Maysson Al-Haj Ibrahim, Sabah Jassim, Michael Anthony Cawthorne, and Kenneth Langlands. A Topology-Based Score for Pathway Enrichment, Journal of Computational Biology. May 2012, 19(5): 563-573
}
\author{
Ivana Ihnatova
}
\keyword{htest}
