% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{res}
\alias{res}
\title{Functions to extract and display results}
\usage{
res(object)
}
\arguments{
\item{object, x}{an output from one of following function \code{"SPIA","PRS", "CePA", "PWEA", "TAPPA", "TopologyGSA","DEGraph", "clipper" }}

\item{...}{other arguments}
}
\description{
Functions to extract and display results
}
