% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAPPA.R
\name{TAPPA}
\alias{TAPPA}
\title{Topological Analysis of Pathway Phenotype Association (TAPPA)}
\usage{
TAPPA(x, group, pathways, type, which = "proteins", edgeType = NULL,
  preparePaths = TRUE, norm.method = NULL, test.method = NULL,
  test = t.test, normalize = TRUE, verbose = FALSE,
  both.directions = TRUE, maxNodes = 150, minEdges = 0,
  commonTh = 2, filterSPIA = FALSE, convertTo = "none",
  convertBy = NULL)
}
\arguments{
\item{x}{An \code{ExpressionSet} object or a gene expression data matrix or count matrix, rows refer to genes, columns to samples}

\item{group}{Name or number of the phenoData column or a character vector or factor that contains required class assigments}

\item{pathways}{A list of pathways in a form from \code{graphite} package or created by \code{preparePathways()}}

\item{type}{Type of the input data, \code{"MA"} for microarray and \code{"RNASeq"} for RNA-Seq}

\item{which}{Character, which type of nodes is preserved in a pathway. Possible values are \code{"proteins"},\code{"metabolites"},\code{"mixed"}}

\item{edgeType}{Character, which type of edges is preserved in a pathway. If \code{NULL}, all edges are kept.}

\item{preparePaths}{Logical, by default the pathways are transformed with \code{preparePathways()}. Use \code{FALSE}, if you have done this transformation separately}

\item{norm.method}{Character, the method to normalize RNAseq data. If \code{NULL} then vst-normalization is performed. Possible values are: \code{"edgeR", "vst", "rLog", "none"}}

\item{test.method}{Character, the method for differentiall expression analysis of RNAseq data. If \code{NULL} then \code{"voomlimma"} is used. Possible values are: \code{"DESeq2", "voomlimma", "vstlimma", "edgeR"}. This analysis is needed only for the visualization.}

\item{test}{Function implementing a statistical test comparing PCI scores between groups. It is employed as \code{test(PCI~group)$p.value}, where \code{PCI} is a numeric vector of the same length as \code{group}}

\item{normalize}{Logical, should data be normalized?}

\item{verbose}{Logical, if \code{TRUE} names of the pathways are printed as they are analysed}

\item{both.directions, maxNodes, minEdges, commonTh, filterSPIA, convertTo, convertBy}{Arguments for the \code{preparePathways()}}
}
\value{
A list,
\item{res}{A data frame, rows refer to pathways. Columns contain: number of valid PCI-scores, median, min and max of the PCI scores for each group of samples, p-value of the \code{test} (\code{p.val}) and adjusted p-value (\code{p.adj}). If less than two nodes are present in the data, the function puts \code{NA}'s in all columns. }
\item{topo.sig}{\code{NULL}, it is preserved for the compatibility with other methods implemented in this package}
\item{degtest}{A numeric vector of gene-level differential expression statistics}
}
\description{
The functions analyses the differential expression of pathways via TAPPA method. Expression is compared between two groups of samples by Mann-Whitney test. P-values are later adjusted for multiple hypothesis testing by Benjamini-Hochberg's FDR method.
}
\examples{
if (require(breastCancerVDX)) {
data("vdx")
pathways<-pathways("hsapiens","biocarta")[1:10]
MAdata<-Biobase::exprs(vdx)[,1:10]
rownames(MAdata)<-Biobase::fData(vdx)[,"Gene.symbol"]
MAdata<-MAdata[!duplicated(rownames(MAdata)),]

TAPPA(MAdata, Biobase::pData(vdx)[,"er"][1:10], pathways, type="MA", convertTo="SYMBOL")
}
}
\references{
Gao, S. and Wang, X. (2007) TAPPA: topological analysis of pathway phenotype association. Bioinformatics, 23, pages 3100-3102
}
\author{
Ivana Ihnatova
}
\keyword{htest}
