% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopologyGSA.r
\name{TopologyGSA}
\alias{TopologyGSA}
\title{Gene set analysis exploiting the topology of a pathway (TopologyGSA)}
\usage{
TopologyGSA(x, group, pathways, type, which = "proteins",
  edgeType = NULL, preparePaths = TRUE, norm.method = NULL,
  test.method = NULL, method = "mean", nperm = 1000, alpha = 0.05,
  testCliques = FALSE, both.directions = TRUE, maxNodes = 150,
  minEdges = 0, commonTh = 2, filterSPIA = FALSE,
  convertTo = "none", convertBy = NULL)
}
\arguments{
\item{x}{An \code{ExpressionSet} object or a gene expression data matrix or count matrix, rows refer to genes, columns to samples}

\item{group}{Name or number of the phenoData column or a character vector or factor that contains required class assigments}

\item{pathways}{A list of pathways in a form from \code{graphite} package or created by \code{preparePathways()}}

\item{type}{Type of the input data, \code{"MA"} for microarray and \code{"RNASeq"} for RNA-Seq}

\item{which}{Character, which type of nodes is preserved in a pathway. Possible values are \code{"proteins"},\code{"metabolites"},\code{"mixed"}}

\item{edgeType}{Character, which type of edges is preserved in a pathway. If \code{NULL}, all edges are kept.}

\item{preparePaths}{Logical, by default the pathways are transformed with \code{preparePathways()}. Use \code{FALSE}, if you have done this transformation separately}

\item{norm.method}{Character, the method to normalize RNAseq data. If \code{NULL} then vst-normalization is performed. Possible values are: \code{"edgeR", "vst", "rLog", "none"}}

\item{test.method}{Character, the method for differentiall expression analysis of RNAseq data. If \code{NULL} then \code{"voomlimma"} is used. Possible values are: \code{"DESeq2", "voomlimma", "vstlimma", "edgeR"}. This analysis is needed only for the visualization.}

\item{method}{Either \code{"var"} and \code{"mean"}. Determine the type of test used by topologyGSA.}

\item{nperm}{Numeric, number of permutations.}

\item{alpha}{Numeric, threshold for statistical significance of variance test. It influences the method for the mean test}

\item{testCliques}{Logical, if \code{TRUE}, then the test is also performed on individual cliques. It can be very computationally complex.}

\item{both.directions, maxNodes, minEdges, commonTh, filterSPIA, convertTo, convertBy}{Arguments for the \code{preparePathways()}}
}
\value{
A list
\item{res}{a list with one entry for each successfully analyzed pathway }
\item{topo.sig}{if \code{testCliques=TRUE}, a list where each slot contains the pvalues and a list of cliques in one pathway. \code{NULL} otherwise}
\item{degtest}{A numeric vector of gene-level differential expression statistics}
}
\description{
TopologyGSA method uses graphical models to test the differential expression of a pathway. It also highlights pathway components involved in the deregulation.
}
\details{
The method requires a Directed Acyclic Graph (DAG). Therefore if a pathway contain also undirected or bidirected edges and error is thrown.

The user can further specify for the mean test:
\enumerate{
 \item \strong{perms} number of permutations of the test,
 \item \strong{paired} logical, if TRUE Hotelling test for paired samples is calculated and the test on the variances is not performed
}

Or for the variance test:
\enumerate{
\item \strong{variance} logical, if TRUE the estimates of the covariance matrices are included in the result.
\item \strong{s1} First group covariance matrix estimation.
\item \strong{s2} Second group covariance matrix estimation.
}
}
\examples{
\dontrun{
if (require(breastCancerVDX)) {
data("vdx")
pathways<-pathways("hsapiens","biocarta")[1:3]
MAdata<-Biobase::exprs(vdx)[,1:10]
rownames(MAdata)<-Biobase::fData(vdx)[,"Gene.symbol"]
MAdata<-MAdata[!duplicated(rownames(MAdata)),]

TopologyGSA(MAdata, Biobase::pData(vdx)[,"er"][1:10], pathways, type="MA", convertTo="SYMBOL", nperm=10)
}
}
}
\references{
Massa MS, Chiogna M, Romualdi C. Gene set analysis exploiting the topology of a pathway. BMC System Biol. 2010 Sep 1;4:121.
}
\author{
Ivana Ihnatova
}
\keyword{htest}
