% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIdentifiersByVector.R
\name{convertIdentifiersByVector}
\alias{convertIdentifiersByVector}
\title{Convert pathway identifiers}
\usage{
convertIdentifiersByVector(pathway, conv.table)
}
\arguments{
\item{pathway}{An object of class \code{Pathway}}

\item{conv.table}{A data.frame, in which the first column contains the type
and the identifiers present in the pathway separeated by \code{:}
and the second column contains the new identifiers and
the third columns contains the types of the new identifiers}
}
\value{
A \code{Pathway} with new identifiers of the nodes
}
\description{
This function converts identifiers of nodes in a pathway. It uses the user specified named vector for the conversion.
}
\examples{
g<-pathways("hsapiens","kegg")
ng<-sapply(g, function(x) length(nodes(x,"mixed")))
g<-g[[which.min(ng)]]
conv<-data.frame(orig=nodes(g,"mixed"), new=LETTERS[seq_len(min(ng))],newtype=rep("LETTERS",min(ng)))
gc<-convertIdentifiersByVector(g, conv.table = conv)@protEdges
}
\seealso{
\code{\link[graphite]{Pathway-class}}
}
\author{
Ivana Ihnatova
}
\keyword{manip}
