\name{res}
\alias{res}
\alias{topo.sig}
\alias{degtable}
\title{
Function to extract parts of object
}
\description{
Function extracts part of an object named \code{"res"}, \code{"topo.sig"}, \code{"degtable"}
}
\usage{
res(object)
topo.sig(object)
degtable(object)
}

\arguments{
  \item{object}{
Object of defined class. Methods for \code{topResult} are available in this package
}
}

\value{
Extracted parts of an object. Data type varies between parts and the origin of the object
}
\author{
Ivana Ihnatova
}
\keyword{ manip }

